requirements.txt
requests
beautifulsoup4
python-whois
flask
flask-cors

